requirements.txt
streamlit
google-generativeai
python-dotenv
pandas
faiss-cpu
scikit-learn
